+++ addons/members-subscriptions/src/functions-products.php
@@ -21,6 +21,20 @@ function init_product_hooks() {
 init_product_hooks();
 
 /**
+ * Get subscription period options for dropdowns
+ * 
+ * @return array Array of period options
+ */
+function get_subscription_period_options() {
+    return [
+        'day'   => __('Day(s)', 'members'),
+        'week'  => __('Week(s)', 'members'),
+        'month' => __('Month(s)', 'members'),
+        'year'  => __('Year(s)', 'members'),
+    ];
+}
+
+/**
  * Process subscription form submission
  */
 function process_subscription_form() {